requirements (2).txt
python-telegram-bot==20.7
fastapi
uvicorn
rembg
pillow
onnxruntime
